+++ test/fixture/mock-generator/readonly-properties-ad-hoc/builder.php
@@ -1,0 +1,10 @@
+<?php
+
+$builder = $factory->create(
+    [
+        'readonly string propertyA' => null,
+        'var readonly ?int propertyB' => null,
+    ]
+);
+
+return $builder->named('MockGeneratorReadonlyPropertiesAdHoc');