+++ test/fixture/mock-generator/readonly-properties-ad-hoc/expected.php
@@ -1,0 +1,13 @@
+<?php
+
+class MockGeneratorReadonlyPropertiesAdHoc
+implements \Eloquent\Phony\Mock\Mock
+{
+    public readonly string $propertyA;
+    public readonly ?int $propertyB;
+    private static $_uncallableMethods = [];
+    private static $_traitMethods = [];
+    private static $_customMethods = [];
+    private static $_staticHandle;
+    private $_handle;
+}